refactor: colocate quest template service tests
